[MOB-3152] Introduce EcosiaSemanticColors into color palette
